requirements.txt
streamlit==1.38.0
tensorflow==2.17.0
pillow==10.4.0
numpy==1.26.4
scikit-image==0.24.0
scikit-learn==1.5.1
streamlit-drawable-canvas==0.9.3
matplotlib==3.9.2